+++ src/machinable/query.py
@@ -5,6 +5,12 @@ from machinable.types import Optional, Union, VersionType
 
 
 class Query:
+    def by_id(self, uuid: str) -> Optional[Interface]:
+        return Interface.find_by_id(uuid)
+
+    def from_directory(self, directory: str) -> Interface:
+        return Interface.from_directory(directory)
+
     def __call__(
         self,
         module: Union[str, Interface, None] = None,