+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 # Unreleased
 
 - Use nanoseconds since epoch in timestamp
+- Adds `get.from_directory` and `get.by_id` to query
 
 # v4.8.4